Module lay_rs::engine

Contains all the logic for running the animations, execute the scheduled changes to the models, and prepare the render tree The scene is drawn in 4 stages:

The layout step calculates the dimensions and position of a node and generates a transformation Matrix

The draw step generates a displaylist

The render step uses the displaylist to generate a texture of the node

The compose step generates the final image using the textures

The LayersEngine is the main engine responsible for managing layers and rendering.

Usage:

let engine = LayersEngine::new();
engine.add_layer(Layer::new());
engine.render();

Modules

animation

Timing functions for animations:

Structs

AnimatedNodeChange

AnimationRef

AnimationState

A struct that contains the state of a given animation.

HandlerRef

LayersEngine

Public API for the Layers Engine

NodeRef

An identifier for a node in the three storage

PointerCallback

SceneNode

Timestamp

TransactionCallback

TransactionRef

Enums

PointerEventType

TransactionEventType

Traits

Command

A trait for objects that can be exectuded by the engine.

SyncCommand