Module lay_rs::prelude::taffy

Modules

compute

Low-level access to the layout algorithms themselves. For a higher-level API, see the TaffyTree struct.

geometry

Geometric primitives useful for layout

prelude

Commonly used types

style

A typed representation of CSS style properties in Rust. Used as input to layout computation.

style_helpers

Helper functions which it make it easier to create instances of types in the style and geometry modules.

tree

Contains both a high-level interface to Taffy using a ready-made node tree, and a set of traits for defining custom node trees.

util

Helpful misc. utilities such as a function to debug print a tree

Structs

Cache

A cache for caching the results of a sizing a Grid Item or Flexbox Item

CollapsibleMarginSet

A set of margins that are available for collapsing with for block layout’s margin collapsing

Layout

The final result of a layout algorithm for a single node.

LayoutInput

A struct containing the inputs constraints/hints for laying out a node, which are passed in by the parent

LayoutOutput

A struct containing the result of laying a single node, which is returned up to the parent node

Line

An abstract “line”. Represents any type that has a start and an end

MinMax

Generic struct which holds a “min” value and a “max” value

NodeId

A type representing the id of a single node in a tree of nodes

Point

A 2-dimensional coordinate.

Rect

An axis-aligned UI rectangle

Size

The width and height of a Rect

Style

A typed representation of the CSS style information for a single node.

TaffyTree

An entire tree of UI nodes. The entry point to Taffy’s high-level API.

Enums

AbsoluteAxis

The simple absolute horizontal and vertical axis

AbstractAxis

The CSS abstract axis https://www.w3.org/TR/css-writing-modes-3/#abstract-axes

AlignContent

Sets the distribution of space between and around content items For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

AlignItems

Used to control how child nodes are aligned. For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

AvailableSpace

The amount of space available to a node in a given axis https://www.w3.org/TR/css-sizing-3/#available

Dimension

A unit of linear measurement

Display

Sets the layout used for the children of this node

FlexDirection

The direction of the flexbox layout main axis.

FlexWrap

Controls whether flex items are forced onto one line or can wrap onto multiple lines.

GridAutoFlow

Controls whether grid items are placed row-wise or column-wise. And whether the sparse or dense packing algorithm is used.

GridTrackRepetition

The first argument to a repeated track definition. This type represents the type of automatic repetition to perform.

LengthPercentage

A unit of linear measurement

LengthPercentageAuto

A unit of linear measurement

MaxTrackSizingFunction

Maximum track sizing function

MinTrackSizingFunction

Minimum track sizing function

Overflow

How children overflowing their container should affect layout

Position

The positioning strategy for this item.

RequestedAxis

An axis that layout algorithms can be requested to compute a size for

RunMode

Whether we are performing a full layout, or we merely need to size the node

SizingMode

Whether styles should be taken into account when computing size

TaffyError

An error that occurs while trying to access or modify a node’s children by index.

TrackSizingFunction

The sizing function for a grid track (row/column) See https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-columns

Traits

FromFlex

Trait to create constant percent values from plain numbers

FromLength

Trait to create absolute length values from plain numbers

FromPercent

Trait to create constant percent values from plain numbers

LayoutPartialTree

Any type that implements LayoutPartialTree can be laid out using Taffy’s algorithms

MaybeMath

A trait to conveniently calculate minimums and maximums when some data may not be defined

MaybeResolve

Trait to encapsulate behaviour where we need to resolve from a potentially context-dependent size or dimension into a context-independent size or dimension.

PrintTree

Trait used by the print_tree method which prints a debug representation

ResolveOrZero

Trait to encapsulate behaviour where we need to resolve from a potentially context-dependent size or dimension into a context-independent size or dimension.

RoundTree

Trait used by the round_layout method which takes a tree of unrounded float-valued layouts and performs rounding to snap the values to the pixel grid.

TaffyAuto

Trait to abstract over auto values

TaffyFitContent

Trait to create fit-content(…) values from plain numbers

TaffyGridLine

Trait to abstract over grid line values

TaffyGridSpan

Trait to abstract over grid span values

TaffyMaxContent

Trait to abstract over max_content values

TaffyMinContent

Trait to abstract over min_content values

TaffyZero

Trait to abstract over zero values

TraversePartialTree

This trait is Taffy’s abstraction for downward tree traversal. However, this trait does not require access to any node’s other than a single container node’s immediate children unless you also intend to implement TraverseTree.

TraverseTree

A marker trait which extends TraversePartialTree with the additional guarantee that the child/children methods can be used to recurse infinitely down the tree. Is required by the RoundTree and the PrintTree traits.

Functions

auto

Returns the auto value for that type

compute_block_layout

Computes the layout of LayoutPartialTree according to the block layout algorithm

compute_cached_layout

Attempts to find a cached layout for the specified node and layout inputs.

compute_flexbox_layout

Computes the layout of LayoutPartialTree according to the flexbox algorithm

compute_grid_layout

Grid layout algorithm This consists of a few phases:

compute_hidden_layout

Creates a layout for this node and its children, recursively. Each hidden node has zero size and is placed at the origin

compute_leaf_layout

Compute the size of a leaf node (node with no children)

compute_root_layout

Compute layout for the root node in the tree

evenly_sized_tracks

Returns a grid template containing count evenly sized tracks

fit_content

Returns a value of the inferred type which represent a fit-content(…) value with the given argument.

flex

Shorthand for minmax(0, Nfr). Probably what you want if you want exactly evenly sized tracks.

fr

Create a Fraction track sizing function (fr in CSS)

length

Returns a value of the inferred type which represent an absolute length

line

Specifies a grid line to place a grid item between in CSS Grid Line coordinates:

max_content

Returns the auto value for that type

min_content

Returns the auto value for that type

minmax

Returns a MinMax with min value of min and max value of max

percent

Returns a value of the inferred type which represent a percentage

print_tree

Prints a debug representation of the computed layout for a tree of nodes, starting with the passed root node.

repeat

Returns an auto-repeated track definition

round_layout

Rounds the calculated layout to exact pixel values

span

Returns a GridPlacement::Span

zero

Returns the zero value for that type

Type Aliases

AlignSelf

Used to control how the specified nodes is aligned. Overrides the parent Node’s AlignItems property. For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

GridPlacement

A grid line placement specification. Used for grid-[row/column]-[start/end]. Named tracks are not implemented.

JustifyContent

Sets the distribution of space between and around content items For Flexbox it controls alignment in the main axis For Grid it controls alignment in the inline axis

JustifyItems

Used to control how child nodes are aligned. Does not apply to Flexbox, and will be ignored if specified on a flex container For Grid it controls alignment in the inline axis

JustifySelf

Used to control how the specified nodes is aligned. Overrides the parent Node’s JustifyItems property. Does not apply to Flexbox, and will be ignored if specified on a flex child For Grid it controls alignment in the inline axis

NonRepeatedTrackSizingFunction

The sizing function for a grid track (row/column) (either auto-track or template track) May either be a MinMax variant which specifies separate values for the min-/max- track sizing functions or a scalar value which applies to both track sizing functions.

TaffyResult

The error Taffy generates on invalid operations