Struct lay_rs::prelude::taffy::geometry::Rect

pub struct Rect<T> {
    pub left: T,
    pub right: T,
    pub top: T,
    pub bottom: T,
}

An axis-aligned UI rectangle

Fields
left: T

This can represent either the x-coordinate of the starting edge, or the amount of padding on the starting side.

The starting edge is the left edge when working with LTR text, and the right edge when working with RTL text.

right: T

This can represent either the x-coordinate of the ending edge, or the amount of padding on the ending side.

The ending edge is the right edge when working with LTR text, and the left edge when working with RTL text.

top: T

This can represent either the y-coordinate of the top edge, or the amount of padding on the top side.

bottom: T

This can represent either the y-coordinate of the bottom edge, or the amount of padding on the bottom side.

Implementations

impl<T> Rect<T>
where T: Add,

pub fn grid_axis_sum(self, axis: AbsoluteAxis) -> <T as Add>::Output

Get either the width or height depending on the AbsoluteAxis passed in

impl<T> Rect<T>

pub fn map<R, F>(self, f: F) -> Rect<R>
where F: Fn(T) -> R,

Applies the function f to the left, right, top, and bottom properties

This is used to transform a Rect<T> into a Rect<R>.

pub fn horizontal_components(self) -> Line<T>

Returns a Line<T> representing the left and right properties of the Rect

pub fn vertical_components(self) -> Line<T>

Returns a Line<T> containing the top and bottom properties of the Rect

impl Rect<f32>

pub const ZERO: Rect<f32> = _

Creates a new Rect with 0.0 as all parameters

pub const fn new(start: f32, end: f32, top: f32, bottom: f32) -> Rect<f32>

Creates a new Rect

impl Rect<Dimension>

pub const fn from_length( start: f32, end: f32, top: f32, bottom: f32, ) -> Rect<Dimension>

Create a new Rect with Dimension::Length

pub const fn from_percent( start: f32, end: f32, top: f32, bottom: f32, ) -> Rect<Dimension>

Create a new Rect with Dimension::Percent

impl<T> Rect<T>
where T: TaffyZero,

pub const fn zero() -> Rect<T>

Returns a Rect where the left, right, top, and bottom values are all the zero value of the contained type (e.g. 0.0, Some(0.0), or Dimension::Length(0.0))

impl<T> Rect<T>
where T: TaffyAuto,

pub const fn auto() -> Rect<T>

Returns a Rect where the left, right, top, and bottom values are all the auto value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T> Rect<T>
where T: TaffyMinContent,

pub const fn min_content() -> Rect<T>

Returns a Rect where the left, right, top, and bottom values are all the min_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T> Rect<T>
where T: TaffyMaxContent,

pub const fn max_content() -> Rect<T>

Returns a Rect where the left, right, top, and bottom values are all the max_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T> Rect<T>
where T: TaffyFitContent,

pub fn fit_content(argument: LengthPercentage) -> Rect<T>

Returns a Rect where the left, right, top and bottom values are all constant fit_content value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Length(2.1))

impl<T> Rect<T>
where T: FromLength,

pub fn length<Input>(value: Input) -> Rect<T>
where Input: Into<f32> + Copy,

Returns a Rect where the left, right, top and bottom values are all the same given absolute length

impl<T> Rect<T>
where T: FromPercent,

pub fn percent<Input>(percent_value: Input) -> Rect<T>
where Input: Into<f32> + Copy,

Returns a Rect where the left, right, top and bottom values are all constant percent value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Length(2.1))