Trait lay_rs::prelude::taffy::ResolveOrZero

pub trait ResolveOrZero<TContext, TOutput>
where
    TOutput: TaffyZero,
{
    // Required method
    fn resolve_or_zero(self, context: TContext) -> TOutput;
}

Trait to encapsulate behaviour where we need to resolve from a potentially context-dependent size or dimension into a context-independent size or dimension.

Will return a default value if it unable to resolve.

Required Methods

fn resolve_or_zero(self, context: TContext) -> TOutput

Resolve a dimension that might be dependent on a context, with a default fallback value