Module lay_rs::types

Types used in the library to describe Layers properties

Structs

BorderRadius

BorderRadiusDiff

Color

GradientLinear

GradientRadial

M44

Matrix

Point

Point3d

Rectangle

Size

V3

Enums

BlendMode

BorderStyle

PaintColor

Type Aliases

Image

Image describes a two dimensional array of pixels to draw. The pixels may be decoded in a raster bitmap, encoded in a Picture or compressed data stream, or located in GPU memory as a GPU texture.

Picture