Trait lay_rs::view::BuildLayerTree

pub trait BuildLayerTree {
    // Required method
    fn build_layer_tree_internal(
        &self,
        viewlayer_tree: &LayerTree,
        cache_viewlayer: &mut HashMap<String, VecDeque<NodeRef>>,
    );

    // Provided method
    fn build_layer_tree(&self, viewlayer_tree: impl AsRef<LayerTree>) { ... }
}

A trait for structs that accept a layertree. It is implemented for Layer it generates and updates the properties of a hierarchy of layers described by the layertree

Required Methods

fn build_layer_tree_internal( &self, viewlayer_tree: &LayerTree, cache_viewlayer: &mut HashMap<String, VecDeque<NodeRef>>, )

Provided Methods

fn build_layer_tree(&self, viewlayer_tree: impl AsRef<LayerTree>)

Object Safety

This trait is not object safe.